move `migrateStylesheet` into `codemods/css/migrate.ts`
