+++ packages/@tailwindcss-upgrade/src/index.test.ts
@@ -4,8 +4,8 @@ import path from 'node:path'
 import postcss from 'postcss'
 import { expect, it } from 'vitest'
 import { formatNodes } from './codemods/css/format-nodes'
+import { migrateContents } from './codemods/css/migrate'
 import { sortBuckets } from './codemods/css/sort-buckets'
-import { migrateContents } from './migrate'
 
 const css = dedent
