+++ packages/@tailwindcss-upgrade/src/index.ts
@@ -5,6 +5,7 @@ import fs from 'node:fs/promises'
 import path from 'node:path'
 import postcss from 'postcss'
 import { formatNodes } from './codemods/css/format-nodes'
+import { migrate as migrateStylesheet } from './codemods/css/migrate'
 import { sortBuckets } from './codemods/css/sort-buckets'
 import { migrate as migrateTemplate } from './codemods/template/migrate'
 import { prepareConfig } from './codemods/template/prepare-config'
@@ -12,7 +13,6 @@ import { help } from './commands/help'
 import {
   analyze as analyzeStylesheets,
   linkConfigs as linkConfigsToStylesheets,
-  migrate as migrateStylesheet,
   split as splitStylesheets,
 } from './migrate'
 import { migrateJsConfig } from './migrate-js-config'
